перенес типы и роутер в отдельную папку, убрал консоль лог
