--- source/app/Router.ts
@@ -1,12 +1,0 @@
-export class Router{
-    routes = []
-    root = '/'
-    constructor () {
-        this.routes = []
-        this.root = '/';
-        
-
-    }
-    
-
-}
