+++ source/shared/Router/RouterTypes.ts
@@ -1,0 +1,12 @@
+import { View } from "@/app/View";
+
+interface configRoute{
+    path?: RegExp;
+    view?: View;
+}
+
+export interface Routes{
+    paths?: (
+        configRoute[]
+    )
+}
